NF: split some tests into foreground and background
The only reason for this split is that it was seen that in foreground, those test fails. It is not yet known why
