+++ AnkiDroid/src/test/java/com/ichi2/anki/dialogs/RecursivePictureMenuForegroundTest.java
@@ -16,16 +16,10 @@
 
 package com.ichi2.anki.dialogs;
 
-import android.view.View;
-
 import com.ichi2.anki.R;
 import com.ichi2.anki.RobolectricTest;
-import com.ichi2.anki.analytics.UsageAnalytics;
-import com.ichi2.anki.dialogs.HelpDialog.LinkItem;
 import com.ichi2.anki.dialogs.RecursivePictureMenu.Item;
 import com.ichi2.anki.dialogs.RecursivePictureMenu.ItemHeader;
-import com.ichi2.anki.dialogs.utils.FragmentTestActivity;
-import com.ichi2.anki.dialogs.utils.RecursivePictureMenuUtil;
 import com.ichi2.utils.ArrayUtil;
 
 import org.junit.Ignore;
@@ -33,46 +27,17 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 
 import androidx.annotation.NonNull;
-import androidx.recyclerview.widget.RecyclerView;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import static com.ichi2.testutils.AnkiAssert.assertDoesNotThrow;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.hasSize;
-import static org.hamcrest.Matchers.is;
 
 @RunWith(AndroidJUnit4.class)
-public class RecursivePictureMenuTest extends RobolectricTest {
+public class RecursivePictureMenuForegroundTest extends RobolectricTest {
     private final @NonNull RecursivePictureMenuDelegate mDelegate = new RecursivePictureMenuDelegate(this);
-
-    @Test
-    public void testNormalStartupSelectingItem() {
-        Item linkedItem = mDelegate.getItemLinkingTo(R.string.link_anki);
-
-        RecyclerView v = mDelegate.getRecyclerViewFor(linkedItem);
-        mDelegate.clickChildAtIndex(v, 0);
-
-        assertThat(mDelegate.getActivity().getLastUrlOpened(), is(getResourceString(R.string.link_anki)));
-    }
-
-
-    @Test
-    public void testSelectingHeader() {
-        int numberOfChildItems = 2;
-        Item header = mDelegate.getHeaderWithSubItems(numberOfChildItems);
-        RecyclerView v = mDelegate.getRecyclerViewFor(header);
-        mDelegate.clickChildAtIndex(v, 0);
-
-        RecursivePictureMenu currentMenu = (RecursivePictureMenu) mDelegate.getActivity().getLastShownDialogFragment();
-
-        RecyclerView rv =  RecursivePictureMenuUtil.getRecyclerViewFor(currentMenu);
-
-        assertThat("Unexpected number of items - check the adapter", rv.getChildCount(), is(numberOfChildItems));
-    }
-
 
     @Test
     @Ignore("Not implemented")
